ContrastBase in lib-builtin-effects
