+++ src/effects/Contrast.h
@@ -10,48 +10,18 @@
 #define __AUDACITY_CONTRAST_DIALOG__
 
 #include "wxPanelWrapper.h" // to inherit
+#include "ContrastBase.h"
 
 class wxButton;
 class wxTextCtrl;
 
-class Envelope;
 class NumericTextCtrl;
-class WaveTrack;
 
 //----------------------------------------------------------------------------
 // ContrastDialog
 //----------------------------------------------------------------------------
 
 // Declare window functions
-
-class ContrastBase
-{
-public:
-   double mT0;
-   double mT1;
-   double mProjectRate;
-   double mStartTimeF;
-   double mEndTimeF;
-   double mStartTimeB;
-   double mEndTimeB;
-
-protected:
-   float foregrounddB;
-   float backgrounddB;
-   bool mForegroundIsDefined;
-   bool mBackgroundIsDefined;
-   double mT0orig;
-   double mT1orig;
-
-   bool mDoBackground;
-   bool GetDB(float& dB);
-   void SetStartAndEndTime();
-
-   double length;
-
-private:
-   virtual AudacityProject& GetProject() = 0;
-};
 
 class ContrastDialog final : public ContrastBase, public wxDialogWrapper
 {
